chore: provide jenkins setup
This setup uses nginx as a reverse proxy for accessing Jenkins, which is running on port 8090. It
will be possible to add other servers to this nginx setup as and when required.

Jenkins runs as a service on the host (as opposed to a container) because I want to be able to run
jobs that access and write files on the server. I obviously can't use a build agent for those jobs.
Later on, I may add my Raspberry Pis as build agents.

However, the nginx setup runs as a container just because that setup seems quite easy to manage.